moved nameserver error to utility
